~ moved testing of `#combine_paths` tests into **test/unit/util/tc_combine_paths.rb**
